# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Gerdt_LR1",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Assignments/show-assigment/{id}": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            }
          }
        }
      }
    },
    "/api/Assignments/user-assigments": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "solved",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              }
            }
          }
        }
      }
    },
    "/api/Assignments/{id}": {
      "delete": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Assignments/{id}/question-answer": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Assignments/{id}/switch-direction": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Assignments/create-assignment": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignForMeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignForMeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignForMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Assignments/generate": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAssignmentsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAssignmentsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAssignmentsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerDto": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AssignForMeDto": {
        "type": "object",
        "properties": {
          "termId": {
            "type": "integer",
            "format": "int32"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          }
        },
        "additionalProperties": false
      },
      "Assignment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "termId": {
            "type": "integer",
            "format": "int32"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          }
        },
        "additionalProperties": false
      },
      "Direction": {
        "enum": [
          "EnToRu",
          "RuToEn"
        ],
        "type": "string"
      },
      "GenerateAssignmentsDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}